work fine added operation in db

+++ src/main/java/ru/smoke/skill/finalproject/v1/service/EmployeeService.java
@@ -2,11 +2,12 @@ package ru.smoke.skill.finalproject.v1.service;
 
 import ru.smoke.skill.finalproject.v1.model.Employee;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
 public interface EmployeeService {
     public List<Employee> getAllEmployees();
-    public Integer getBalance(Integer id);
+    public BigDecimal getBalance(Long id);
     public void saveEmployee(Employee employee);
 }
